+++ R/boxplot2.R
@@ -1,8 +1,4 @@
 # $Id: boxplot2.R 2052 2015-06-02 19:03:20Z warnes $
-boxplot.n <- function( ..., top=FALSE, shrink=1.0, textcolor=NULL )
-    {
-        .Defunct("boxplot2", package="gplots")
-    }
 
 boxplot2  <- function( ..., top=FALSE, shrink=1.0, textcolor=NULL )
   {