+++ R/lmplot2.R
@@ -1,25 +1,3 @@
-plot.lm2 <- function(
-                     x,
-                     which = 1:5,
-                     caption = c("Residuals vs Fitted", "Normal Q-Q plot",
-                                 "Scale-Location plot", "Cook's distance plot"),
-                     panel = panel.smooth,
-                     sub.caption = deparse(x$call),
-                     main = "",
-                     ask,
-                     ...,
-                     id.n = 3,
-                     labels.id = names(residuals(x)),
-                     cex.id = 0.75,
-                     band=TRUE,
-                     rug=TRUE,
-                     width=1/10,
-                     max.n=5000
-                     )
-{
-  .Defunct("lmplot2", "gplots")
-}
-
 
 lmplot2 <- function(
                     x,